+++ applications/main/subghz/subghz_last_settings.c
@@ -18,6 +18,7 @@
 #define SUBGHZ_LAST_SETTING_FIELD_FREQUENCY_ANALYZER_FEEDBACK_LEVEL "FeedbackLevel"
 #define SUBGHZ_LAST_SETTING_FIELD_FREQUENCY_ANALYZER_TRIGGER "FATrigger"
 #define SUBGHZ_LAST_SETTING_FIELD_EXTERNAL_MODULE_ENABLED "External"
+#define SUBGHZ_LAST_SETTING_FIELD_EXTERNAL_MODULE_POWER "ExtPower"
 
 SubGhzLastSettings* subghz_last_settings_alloc(void) {
     SubGhzLastSettings* instance = malloc(sizeof(SubGhzLastSettings));
@@ -43,6 +44,7 @@ void subghz_last_settings_load(SubGhzLastSettings* instance, size_t preset_count
     uint32_t temp_frequency_analyzer_feedback_level = 0;
     float temp_frequency_analyzer_trigger = 0;
     bool temp_external_module_enabled = false;
+    bool temp_external_module_power_5v_disable = false;
     //int32_t temp_preset = 0;
     bool frequency_analyzer_feedback_level_was_read = false;
     bool frequency_analyzer_trigger_was_read = false;
@@ -68,6 +70,11 @@ void subghz_last_settings_load(SubGhzLastSettings* instance, size_t preset_count
             fff_data_file,
             SUBGHZ_LAST_SETTING_FIELD_EXTERNAL_MODULE_ENABLED,
             (bool*)&temp_external_module_enabled,
+            1);
+        flipper_format_read_bool(
+            fff_data_file,
+            SUBGHZ_LAST_SETTING_FIELD_EXTERNAL_MODULE_POWER,
+            (bool*)&temp_external_module_power_5v_disable,
             1);
 
     } else {
@@ -99,6 +106,13 @@ void subghz_last_settings_load(SubGhzLastSettings* instance, size_t preset_count
         instance->preset = SUBGHZ_LAST_SETTING_DEFAULT_PRESET;
 
         instance->external_module_enabled = temp_external_module_enabled;
+
+        instance->external_module_power_5v_disable = temp_external_module_power_5v_disable;
+
+        if(instance->external_module_power_5v_disable) {
+            furi_hal_subghz_set_external_power_disable(true);
+            furi_hal_subghz_disable_ext_power();
+        }
 
         // Set selected radio module
         if(instance->external_module_enabled) {
@@ -165,6 +179,13 @@ bool subghz_last_settings_save(SubGhzLastSettings* instance) {
                file,
                SUBGHZ_LAST_SETTING_FIELD_EXTERNAL_MODULE_ENABLED,
                &instance->external_module_enabled,
+               1)) {
+            break;
+        }
+        if(!flipper_format_insert_or_update_bool(
+               file,
+               SUBGHZ_LAST_SETTING_FIELD_EXTERNAL_MODULE_POWER,
+               &instance->external_module_power_5v_disable,
                1)) {
             break;
         }