+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ### New changes
 * SubGHz: New protocols support: CAME Space, Stilmatic / Schellenberg
 * SubGHz: Add Manually - new protocols -> Beninca, Sommer(FSK), IronLogic, DTM Neo, Gibidi, Elmes Electronic (Elmes Poland), CAME Space
-* SubGHz: Remember last external module setting, so if you turn off flipper it will remember last external module setting (only for subghz app)
+* SubGHz: Remember last external module setting and power setting, so if you turn off flipper it will remember last external module settings (only for subghz app)
 * SubGHz: Fix issues when external module is not found but plugins tries to use it, now they will fallback to internal in that case
 * SubGHz: Fixed external CC1101 module power issues, added more checks, fixed issues when launching subghz favourites
 * SubGHz: Removed 330MHz from default freq hopper to make it faster, to change freq hopper settings and remove/add your freqs see -> [Instruction](https://github.com/DarkFlippers/unleashed-firmware/blob/dev/documentation/SubGHzSettings.md)